+++ load.py
@@ -62,11 +62,9 @@ def plugin_app(parent:tk.Widget) -> tk.Frame:
 def journal_entry(cmdr:str, is_beta:bool, system:str, station:str, entry:dict, state:dict) -> None:
     match entry['event']:
         case 'Startup':
-            Context.router.system = system
             Context.router.carrier_state = CarrierStates.Idle
             if Context.route.route != []: Context.route.update_route(0, system)
         case 'FSDJump' | 'Location' | 'SupercruiseExit' if entry.get('StarSystem', system) != Context.router.system:
-            Context.router.system = system
             Context.router.jumped(system, entry)
         case 'CarrierJumpRequest' | 'CarrierLocation' | 'CarrierJumpCancelled':
             Context.router.carrier_event(entry)
@@ -74,8 +72,6 @@ def journal_entry(cmdr:str, is_beta:bool, system:str, station:str, entry:dict, s
             Context.router.set_ship(entry)
         case 'ShipyardSwap':
             Context.router.swap_ship(entry.get('ShipID', ''))
-        case 'Cargo':
-            Context.router.cargo = entry.get('Count', 0)
         case 'SendText':
             if entry.get('Message', '').startswith("!nd "):
                 match entry.get('Message', '')[4:]:
@@ -86,11 +82,14 @@ def journal_entry(cmdr:str, is_beta:bool, system:str, station:str, entry:dict, s
                     case _:
                         copy_to_clipboard(Context.ui.parent, Context.route.next_stop())
 
+    Context.router.system = system
     Context.router.cargo = sum(state.get('Cargo', {}).values())
     Context.ui.update_cargo(Context.router.cargo)
 
 
 def dashboard_entry(cmdr:str, is_beta:bool, entry:dict) -> None:
+    if Context.router:
+        Context.router.dashboard_entry(cmdr, is_beta, entry)
     if Context.overlay:
         Context.overlay.dashboard_entry(cmdr, is_beta, entry)