+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added route icon to indicate neutron star
 * Added waypoint progress to next stop button
 * Now (re-)copies next jump to clipboard on entry to Galaxy Map
+* If a route is cleared once completed the source and destination are reversed
 * Improved release build process to include `virustotal`
 * Updated `github` automated unit tests to exclude slow tests
 * Added `flake8` checks to `github` automated unit tests